+++ influxio/core.py
@@ -3,13 +3,14 @@ from pathlib import Path
 
 from yarl import URL
 
-from influxio.model import InfluxAPI
+from influxio.model import InfluxDbAdapter, SqlAlchemyAdapter
 from influxio.testdata import DataFrameFactory
+from influxio.util.db import get_sqlalchemy_dialects
 
 logger = logging.getLogger(__name__)
 
 
-def copy(source: str, target: str):
+def copy(source: str, target: str, progress: bool = False):
     """
     Copy/transfer data from/to InfluxDB API / InfluxDB line protocol / RDBMS.
 
@@ -17,7 +18,7 @@ def copy(source: str, target: str):
 
     `source` and `target` are resource identifiers in URL format.
 
-    When the InfluxDB API is addressed, the schema is:
+    When InfluxDB is addressed, the schema is:
     http://example:token@localhost:8086/testdrive/demo
 
     This means:
@@ -25,30 +26,47 @@ def copy(source: str, target: str):
     - Authentication: token
     - Bucket: testdrive
     - Measurement: demo
+
+    When an RDBMS is addressed through SQLAlchemy, the schema is:
+    http://username:password@localhost:12345/testdrive/demo
+
+    This means:
+    - Database or schema: testdrive
+    - Table name: demo
     """
-    source = URL(source)
-    target = URL(target)
+    source_url = URL(source)
+    target_url = URL(target)
+
+    sqlalchemy_dialects = get_sqlalchemy_dialects()
 
     logger.info(f"Copying from {source} to {target}")
 
-    if target.scheme.startswith("http"):
-        sink = InfluxAPI.from_url(target)
-    else:
-        raise NotImplementedError(f"Data sink not implemented: {target}")
+    scheme_primary = target_url.scheme.split("+")[0]
 
-    if source.scheme == "testdata":
-        dff = DataFrameFactory(**source.query)
-        df = dff.make(source.host)
+    if target_url.scheme.startswith("http"):
+        sink = InfluxDbAdapter.from_url(target)
+    elif scheme_primary in sqlalchemy_dialects:
+        sink = SqlAlchemyAdapter.from_url(target, progress=True)
+    else:
+        raise NotImplementedError(f"Data sink not implemented: {target_url}")
+
+    if source_url.scheme == "testdata":
+        dff = DataFrameFactory(**source_url.query)
+        df = dff.make(source_url.host)
         sink.write_df(df)
 
-    elif source.scheme == "file":
-        path = Path(source.host).joinpath(Path(source.path).relative_to("/"))
+    elif source_url.scheme == "file":
+        path = Path(source_url.host).joinpath(Path(source_url.path).relative_to("/"))
         # TODO: Determine file type by suffix.
         # TODO: Make `precision` configurable.
-        sink.write_lineprotocol(path)
+        sink.from_lineprotocol(path)
 
-    elif source.scheme.startswith("http"):
-        sink.write_lineprotocol(str(source))
+    elif source_url.scheme.startswith("http"):
+        if isinstance(sink, SqlAlchemyAdapter):
+            source_node = InfluxDbAdapter.from_url(source)
+            sink.write(source_node)
+        else:
+            sink.from_lineprotocol(str(source_url))
 
     else:
-        raise NotImplementedError(f"Data source not implemented: {source}")
+        raise NotImplementedError(f"Data source not implemented: {source_url}")