+++ CHANGES.rst
@@ -10,6 +10,8 @@ in progress
 - Tests: Add test for ``influxio info``
 - Feature: Add reading line protocol format from file
 - Feature: Add reading line protocol format from URL
+- Feature: Export from InfluxDB and import into RDBMS,
+  using SQLAlchemy/pandas/Dask
 
 
 2023-xx-xx 0.1.0